chore: 优化代码结构，将 Add 命令从 Program class 中移动到单独的 class 里
